Split laserpath into two files:laserpath and lasercavity
